libs/githubcrawler/requirements.txt
PyGithub
gitpython
frozendict
